+++ apps/web/src/App.vue
@@ -1,11 +1,12 @@
 <script setup>
-import { ref, computed } from "vue"
-import { sum, subtract } from "@monorepo-series/utils"
+import { ref, computed, onMounted } from "vue"
+import { sum, subtract, UseDate } from "@monorepo-series/utils"
 
 const val1 = ref("")
 const val2 = ref("")
 
 const btnDisabled = computed(() => ![val1, val2].every(val => typeof val.value === "number"))
+const now = ref(UseDate.format(new Date()))
 
 function addValues() {
   const num1 = parseFloat(val1.value)
@@ -39,12 +40,18 @@ async function getFruitFn() {
       alert("获取水果数据失败")
     })
 }
+
+onMounted(() => {
+	setInterval(() => {
+		now.value = UseDate.format(new Date())
+	}, 1000)
+})
 </script>
 
 <template>
   <div class="page-index">
     <h5 class="title">计数器</h5>
-    <div class="app-container">
+    <div class="counter-container">
 			<div class="input-container">
 				<input type="number" v-model="val1" placeholder="请输入val1" />
 				<input type="number" v-model="val2" placeholder="请输入val2" />
@@ -55,27 +62,30 @@ async function getFruitFn() {
         <button :disabled="btnDisabled" @click="subtractValues">相减</button>
       </div>
     </div>
+		<hr>
 		<h5 class="title">接口数据获取</h5>
 		<div>
 			<button id="getFruitBtn" @click="getFruitFn">水果数据</button>
 		</div>
+		<div class="bottom-lock">{{ now }}</div>
   </div>
+
 </template>
 
 <style lang="scss" scoped>
 .page-index {
-	margin: 0 20px;
+	position: relative;
+	min-height: 100vh;
+	border: 10px solid transparent;
 }
 
 .title {
-	font: {
-		size: 1.5rem;
-		font-family: Arial, "Helvetica Neue", Helvetica, sans-serif;
-	}
-	margin: 1rem 0;
+	font-size: 1.5rem;
+	margin: 0 0 10px;
 }
 
-.app-container {
+.counter-container {
+	margin: 0;
 	.input-container {
 		input:nth-child(n+2) {
 			margin-left: 10px;
@@ -89,4 +99,9 @@ async function getFruitFn() {
 	}
 }
 
+.bottom-lock {
+	position: fixed;
+	bottom: 1rem;
+	right: 0.5rem;
+}
 </style>