+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @monorepo-series/web
 
+## 1.0.2
+
+### Patch Changes
+
+- 页面加入当前时间日期，并优化页面展示
+- Updated dependencies
+  - @monorepo-series/utils@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes