+++ demo/play-webpack-1/CHANGELOG.md
@@ -1,5 +1,12 @@
 # play-webpack-1
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @monorepo-series/utils@1.1.0
+
 ## 1.0.1
 
 ### Patch Changes